🚚 Rename: 에러컴포넌트 CustomError 로 변경
